This is the first commit of the gomake project
The purpose of this project is to create a version of the popular GNU
tool make using Go and eventually add some new features including
constant background update on the code and some concurrency in the
evaluation of the commands in the makefile.  In this first iteration
the only thing that exists is a bare outline of what the project will
eventually be.  I'm using a separate structure to try to encode the
different dependencies present in the makefile structure.

TODO:
 - Migrate the parsing out of the main package and into the dependency
   package

 - Once the parsing is correctly building dependencies get the make
   function working

 - After that determine areas where concurrency can be used and how to
   implement the auto update feature

[Fix #1]