Functional infinite make utility
This is a working proof of concept of the continual make utility
written in Go.  It can currently read in a simplified version of a
makefile and based upon those dependencies build a set of files.  It
currently has some pretty hefty limitations, primarily the fact that
it doesn't support variable declarations.  The utility also needs to
go through some major testing among other things.
TODO:

 - Polish the infinite looping functionality to only run over a set of
 files once

 - Add in support for variables in the actual makefile

 - Improve the concurrency model and the filesystem listener

[Fix #5]